+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 - Filter offered resources by catalog
 - Set contract validity period for offered resources
 - Data Consumption: Show all resources instead of catalogs after request
-- Data Offering: show number of agreements per resource
+- Data Offering: Show agreements of resources
 
 ### Fixes
 - Data Consumption: No validation on resource meta data dialog